Extract the subscription registry and subscription record to make Stand more readable; document the methods of the extracted classes and make the classes @Internal.
